Split up useDeviceProvider
